JSON inter-cycle corruption fixer
As we are proceeding with the implementation of a database in the next release, this update introduces a mechanism to check JSON play files for corruption and attempts to repair them between cycles, going off a reference play file. It does not add values, only fixes structure and syntax, particularly in the final logging section where the cutting off has been occuring on rare occasions.